+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - (Enhancement) Add support for ipv4 and ipv6 address pools under subinterface and etherchannel
 - (Enhancement) Add support for `fmc_icmpv6_objects`
+- (Enhancement) Add support for `fmc_device_loopback_interface`
 
 ## 2.0.0-beta3